portfolio-RyanTakkenberg/requirements.txt
azure-core==1.31.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
Click==7.0
cryptography==43.0.1
Flask
idna==3.10
isodate==0.6.1
itsdangerous==1.1.0
jdcal==1.4.1
MarkupSafe==1.1.1
msal==1.31.0
msal-extensions==1.2.0
mysql-connector-python==9.0.0
mysql-connector-python-rf==2.2.2
pandas==2.2.2
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.34

